+++ tests/testthat/test-extract.R
@@ -15,9 +15,11 @@ df8 <- fingertips_data(DomainID = 1938132767, rank = TRUE)
 inds <- c(93081, 93275, 93094)
 df9 <- fingertips_data(inds, ProfileID = 143, AreaTypeID = 3)
 
+# Testing different versions of AreaTypeID = "All"
 df10 <- fingertips_data(IndicatorID = 10101, AreaTypeID = "All")
 df11 <- fingertips_data(DomainID = 1938132954, AreaTypeID = "All")
 df12 <- fingertips_data(ProfileID = 152, AreaTypeID = "All")
+df13 <- fingertips_data(IndicatorID = 10101, ProfileID = 19, AreaTypeID = "All")
 
 ncols <- 26
 
@@ -70,4 +72,5 @@ test_that(paste("number of fields returned by fingertips_data function are", nco
         expect_equal(ncol(df10), ncols)
         expect_equal(ncol(df11), ncols)
         expect_equal(ncol(df12), ncols)
+        expect_equal(ncol(df13), ncols)
 })